Rate limit based on ALB logs and move Lambda CF
* Update the Rate Limit Lambda to check ALB access logs instead of CloudFront access logs
* Move the Lambda CloudFormation into this project (from the Cerberus 'base' CloudFormation stack)
