+++ .gitignore
@@ -48,7 +48,7 @@ hs_err_pid*
 ## Plugin-specific files:
 
 # IntelliJ
-/out/
+out/
 classes
 
 # mpeltonen/sbt-idea plugin
