+++ README.md
@@ -10,7 +10,7 @@ This project contains the serverless components that can be used with Cerberus.
 
 * [Cerberus Health Check](cerberus-health-check-lambda/README.md) - A small serverless API that authenticates and reads a secret from cerberus.
 * [Cerberus Artemis KPI Lambda](cerberus-artemis-kpi-lambda/README.md) - Reference Example function for processing Cerberus metrics from the metrics topic
-* [Cerberus CloudFront Lambda](cerberus-cloudfront-lambda/README.md) - Serverless function for processing CloudFront log events, to enable things such as rate limiting and optionally Google Analytics KPI tracking
+* [Cerberus Rate Limiting Lambda](cerberus-rate-limit-lambda/README.md) - Serverless function for processing CloudFront log events, to enable things such as rate limiting and optionally Google Analytics KPI tracking
 * [Cerberus Lambda VPC](cerberus-lambda-vpc/README.md) - Cloudformation to create a VPC with EIBs and NATs so that Cerberus operators can run lambdas with predictable IP addresses.
 * [Cerberus Metrics Topic](cerberus-metrics-topic/README.md) - Cloudformation to create a SNS topic that is used by various components to publish Cerberus metrics.
 * [Cerberus Admin IAM Role](cerberus-admin-iam-role/README.md) - Creates the admin IAM role to be used by the Lambda serverless components to call the Cerberus Management Service admin endpoints.
