--- cerberus-cloudfront-lambda/build.gradle
@@ -1,28 +1,0 @@
-version '1.1.0'
-
-dependencies {
-    compile group: 'com.amazonaws', name: 'aws-lambda-java-core', version: '1.1.0'
-    compile group: 'com.amazonaws', name: 'aws-lambda-java-log4j', version: '1.0.0'
-    compile group: 'com.amazonaws', name: 'aws-lambda-java-events', version: '1.3.0'
-    compile group: 'com.amazonaws', name: 'aws-java-sdk-cloudformation', version: '1.11.37'
-    compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.8.3'
-    compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.8.3'
-    compile group: 'com.google.guava', name: 'guava', version: '19.0'
-    compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.4'
-    compile group: 'commons-net', name: 'commons-net', version: '3.5'
-    compile group: 'com.amazonaws', name: 'aws-java-sdk-waf', version: '1.11.39'
-    compile group: 'com.fieldju', name: 'slack-client', version: '2.0.0'
-    compile group: 'com.brsanthu', name: 'google-analytics-java', version: '1.1.2'
-
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-    testCompile group: 'org.mockito', name: 'mockito-all', version: '1.9.5'
-}
-
-shadowJar {
-    baseName = 'cerberus-cloudfront-lambda'
-    classifier = null
-}
-
-tasks.assemble.finalizedBy shadowJar
-
-tasks.check.dependsOn jacocoTestReport
