--- cerberus-cloudfront-lambda/src/main/java/com/nike/cerberus/lambda/waf/processor/Processor.java
@@ -1,10 +1,0 @@
-package com.nike.cerberus.lambda.waf.processor;
-
-import com.nike.cerberus.lambda.waf.CloudFrontLogHandlerConfig;
-import com.nike.cerberus.lambda.waf.CloudFrontLogEvent;
-
-import java.util.List;
-
-public interface Processor {
-    void processLogEvents(List<CloudFrontLogEvent> events, CloudFrontLogHandlerConfig params, String bucketName);
-}
