+++ profile/example.properties
@@ -25,3 +25,9 @@ cerberus.lambda_artifact.s3_bucket=[The bucket name to store the compiled lambda
 health_check.regions=[comma deliminated list of regions to deploy the health check lambda to]
 
 cleanup.kms_expiration_period_in_days=[Clean up KMS keys that have been inactive for n days]
+
+rate_limit.blacklist_for_x_mins_after_violation=[The number of minutes to blacklist an IP address for after it has violated the rate limit]
+rate_limit.request_per_minute_limit=[The number of requests per minute allowed per IP address]
+rate_limit.google_analystics_id=[The ID of the Google Analytics account to which the Lambda will send KPIs. This property is nullable]
+rate_limit.slack_web_hook_url=[The Slack HTTP endpoint to which the Lambda will post rate limit notifications. This property is nullable]
+rate_limit.slack_cerberus_icon=[URL of an icon image file to be used in Slack. This property is nullable]
